+++ apps/ingestion/src/controller.ts
@@ -1,8 +1,10 @@
 import type { ControllerDelta, ControllerLong, ControllerMerged, ControllerShort, PilotLong } from "@sr24/types/interface";
 import type { VatsimData } from "@sr24/types/vatsim";
 import { haversineDistance } from "./utils/helpers.js";
+import { getUserRatings } from "./utils/ratings.js";
 import { findPrefixMatch, reduceCallsign } from "./utils/sectors.js";
 
+let cachedLongs: ControllerLong[] = [];
 let cached: ControllerMerged[] = [];
 let updated: ControllerMerged[] = [];
 let added: ControllerMerged[] = [];
@@ -12,6 +14,8 @@ export async function mapControllers(vatsimData: VatsimData, pilotsLong: PilotLo
 		.map((controller) => {
 			if (controller.facility === 0 && !controller.callsign.includes("OBS")) return null;
 			if (controller.frequency === "199.998") return null;
+
+			const cachedController = cachedLongs.find((c) => c.cid === String(controller.cid));
 			return {
 				callsign: controller.callsign,
 				frequency: parseFrequencyToKHz(controller.frequency),
@@ -22,7 +26,7 @@ export async function mapControllers(vatsimData: VatsimData, pilotsLong: PilotLo
 				name: controller.name,
 				server: controller.server,
 				visual_range: controller.visual_range,
-				user_ratings: null,
+				user_ratings: cachedController?.user_ratings || null,
 				logon_time: new Date(controller.logon_time).getTime(),
 				timestamp: new Date(controller.last_updated).getTime(),
 			};
@@ -31,13 +35,14 @@ export async function mapControllers(vatsimData: VatsimData, pilotsLong: PilotLo
 
 	getConnectionsCount(vatsimData, controllersLong, pilotsLong);
 
-	// const userRatings = await getUserRatings(controllersLong.filter((c) => c.user_ratings === null).map((c) => c.cid));
-	// for (const controller of controllersLong) {
-	// 	const ratings = userRatings[controller.cid];
-	// 	if (ratings) {
-	// 		controller.user_ratings = ratings;
-	// 	}
-	// }
+	const uniqueCids = Array.from(new Set(controllersLong.filter((c) => c.user_ratings === null).map((c) => c.cid)));
+	const userRatings = await getUserRatings(uniqueCids);
+	for (const controller of controllersLong) {
+		const ratings = userRatings.get(controller.cid);
+		if (ratings) {
+			controller.user_ratings = ratings;
+		}
+	}
 
 	vatsimData.atis.forEach((atis) => {
 		controllersLong.push({
@@ -58,6 +63,8 @@ export async function mapControllers(vatsimData: VatsimData, pilotsLong: PilotLo
 
 	const merged = await mergeControllers(controllersLong);
 	setControllerDelta(merged);
+
+	cachedLongs = controllersLong;
 
 	return [controllersLong, merged];
 }